# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/healz": {
            "get": {
                "description": "Check health endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "utils"
                ],
                "summary": "healz",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Healz.response"
                        }
                    }
                }
            }
        },
        "/ready": {
            "get": {
                "description": "Check readiness",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "utils"
                ],
                "summary": "ready",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Ready.response"
                        }
                    }
                }
            }
        },
        "/translate/{name}": {
            "post": {
                "description": "Translate input text `fromLang` `toLang`",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "translate something",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Translate to language",
                        "name": "toLang",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Translate from language",
                        "name": "fromLang",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Index",
                        "name": "index",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TranslateResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Healz.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "api.Ready.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Ready"
                }
            }
        },
        "api.TranslateResponse": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string",
                    "example": "test"
                }
            }
        }
    }
}